@@ -12,16 +12,17 @@
     "kepler.gl": "^0.2.2",
     "lodash.debounce": "^4.0.8",
     "lodash.get": "^4.4.2",
-    "react": "^15.6.2",
+    "react": "^16.8.3",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^15.6.1",
+    "react-dom": "^16.8.3",
     "react-palm": "^1.1.2",
     "react-redux": "^4.4.5",
     "react-router": "^3.0.5",
     "react-router-redux": "^4.0.8",
     "react-virtualized": "^9.18.5",
     "redux-actions": "^2.2.1",
-    "redux-thunk": "^1.0.0"
+    "redux-thunk": "^1.0.0",
+    "styled-components": "^4.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
