@@ -110,7 +110,6 @@
     "redux": "^3.0.4",
     "redux-actions": "^2.2.1",
     "reselect": "^3.0.1",
-    "styled-components": "^2.2.4",
     "supercluster": "^3.0.0",
     "type-analyzer": "https://registry.npmjs.org/type-analyzer/-/type-analyzer-0.2.1.tgz",
     "uber-licence": "^3.1.1",
@@ -170,13 +169,14 @@
     "nyc": "^11.4.1",
     "prettier": "^1.7.0",
     "progress-bar-webpack-plugin": "^1.9.3",
-    "react": "^15.3.1",
+    "react": "^16.8.3",
     "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^15.6.1",
+    "react-dom": "^16.8.3",
     "react-hot-loader": "^3.0.0-beta.6",
     "redux-mock-store": "^1.2.1",
     "sinon": "^2.4.1",
     "source-map-loader": "^0.2.1",
+    "styled-components": "^4.1.3",
     "tap-spec": "^4.1.1",
     "tape": "4.8.0",
     "tape-catch": "^1.0.6",
@@ -190,8 +190,9 @@
     "webpack-stats-plugin": "^0.2.1"
   },
   "peerDependencies": {
-    "react": "0.14.x - 16.x",
-    "react-dom": "0.14.x - 16.x"
+    "react": "16.x",
+    "react-dom": "16.x",
+    "styled-components": "^4.x"
   },
   "nyc": {
     "sourceMap": false,
