@@ -32,7 +32,7 @@ export const SelectText = styled.span`
   }
 `;
 
-export const SelectTextBold = SelectText.extend`
+export const SelectTextBold = styled(SelectText)`
   color: ${props => props.theme.textColor};
   font-weight: 500;
 `;
@@ -77,7 +77,7 @@ export const PanelLabelWrapper = styled.div.attrs({
   align-items: self-start;
 `;
 
-export const PanelLabelBold = PanelLabel.extend`
+export const PanelLabelBold = styled(PanelLabel)`
   font-weight: 500;
 `;
 
@@ -242,7 +242,7 @@ export const InputLight = styled.input`
   ${props => props.theme.inputLT}
 `;
 
-export const InlineInput = Input.extend`
+export const InlineInput = styled(Input)`
   ${props => props.theme.inlineInput};
 `;
 
