@@ -218,7 +218,7 @@ const StyledAnimationControls = styled.div`
   }
 `;
 
-const IconButton = Button.extend`
+const IconButton = styled(Button)`
   svg {
     margin: 0 6px;
   }
