@@ -90,7 +90,7 @@ const Tab = styled.div`
 `;
 /* eslint-enable no-unused-vars */
 
-const StyledTitle = CenterFlexbox.extend`
+const StyledTitle = styled(CenterFlexbox)`
   flex-grow: 0;
   color: ${props => props.theme.textColor};
 
