@@ -79,7 +79,7 @@ const StyledPin = styled.div`
   }
 `;
 
-const StyledLayerName = CenterFlexbox.extend`
+const StyledLayerName = styled(CenterFlexbox)`
   color: ${props => props.theme.textColorHl};
   font-size: 12px;
   letter-spacing: 0.43px;
