@@ -55,7 +55,7 @@ const defaultProps = {
   showRemoveLayer: true
 };
 
-const StyledLayerPanelHeader = StyledPanelHeader.extend`
+const StyledLayerPanelHeader = styled(StyledPanelHeader)`
   .layer__remove-layer {
     opacity: 0;
   }
