@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import React from 'react';
+import styled from "styled-components";import React from 'react';
 import classnames from 'classnames';
 import {ArrowDown} from 'components/common/icons';
 import PanelHeaderAction from 'components/side-panel/panel-header-action';
@@ -30,7 +30,7 @@ import {
   StyledPanelHeader
 } from 'components/common/styled-components';
 
-const StyledMapDropdown = StyledPanelHeader.extend`
+const StyledMapDropdown = styled(StyledPanelHeader)`
   height: 48px;
   margin-bottom: 5px;
   opacity: 1;
